fix: moved filter logic to TaskDashboard component.
